+++ src/war_cheats_panel.c
@@ -83,7 +83,7 @@ void createCheatsPanel(WarContext* context)
 
     WarEntity* uiEntity;
 
-    vec2 cheatSize = vec2f(context->originalWindowWidth, 12);
+    vec2 cheatSize = vec2f((f32)context->originalWindowWidth, 12.0f);
     u8Color cheatBackgroundColor = u8RgbaColor(100, 100, 100, 160);
     uiEntity = createUIRect(context, "panelCheat", VEC2_ZERO, cheatSize, cheatBackgroundColor);
     setUIEntityStatus(uiEntity, false);
@@ -165,7 +165,7 @@ void updateCheatsPanel(WarContext* context)
 
         if (wasKeyPressed(input, WAR_KEY_TAB))
         {
-            s32 length = strlen(cheatStatus->text);
+            s32 length = (s32)strlen(cheatStatus->text);
             if (TAB_WIDTH <= STATUS_TEXT_MAX_LENGTH - length)
             {
                 strInsertAt(cheatStatus->text, cheatStatus->position, '\t');
@@ -182,7 +182,7 @@ void updateCheatsPanel(WarContext* context)
         }
         else if (wasKeyPressed(input, WAR_KEY_DELETE))
         {
-            s32 length = strlen(cheatStatus->text);
+            s32 length = (s32)strlen(cheatStatus->text);
             if (cheatStatus->position < length)
             {
                 strRemoveAt(cheatStatus->text, cheatStatus->position);
@@ -190,7 +190,7 @@ void updateCheatsPanel(WarContext* context)
         }
         else if (wasKeyPressed(input, WAR_KEY_RIGHT))
         {
-            s32 length = strlen(cheatStatus->text);
+            s32 length = (s32)strlen(cheatStatus->text);
             if (cheatStatus->position < length)
             {
                 cheatStatus->position++;
@@ -209,7 +209,7 @@ void updateCheatsPanel(WarContext* context)
         }
         else if (wasKeyPressed(input, WAR_KEY_END))
         {
-            s32 length = strlen(cheatStatus->text);
+            s32 length = (s32)strlen(cheatStatus->text);
             cheatStatus->position = length;
         }
 
@@ -223,7 +223,7 @@ void updateCheatsPanel(WarContext* context)
         params.fontSize = cheatText->text.fontSize;
         params.fontData = fontsData[cheatText->text.fontIndex];
 
-        vec2 prefixSize = measureSingleSpriteText("MSG: ", strlen("MSG: "), params);
+        vec2 prefixSize = measureSingleSpriteText("MSG: ", (s32)strlen("MSG: "), params);
         vec2 textSize = measureSingleSpriteText(cheatStatus->text, cheatStatus->position, params);
         cheatCursor->transform.position.x = prefixSize.x + textSize.x;
